Added PVE2 and PVE4 yaml files.
PVE2 represents the state of the rails on April 30, 2011. Later on in the
map Xavier Transit made its way past Woodville/North Gate. There was a
line between Woodville and Spawn as well as rails out east connecting Brom
Harbor and Ravenna to the central area.

PVE4 is mostly complete, but I left out the Brom metro rails and the
stations around Seneca.

Apparently I get sick of rail documentation and take the odd maps off.